+++ R/hyd1d.R
@@ -1,5 +1,4 @@
 #' @name hyd1d
-#' @docType package
 #' @aliases hyd1d-package
 #' 
 #' @title hyd1d - 1d Water Level Interpolation along the Rivers Elbe and Rhine
@@ -43,5 +42,5 @@
 #' @importFrom curl has_internet
 #' @importFrom Rdpack reprompt
 #'
-NULL
+"_PACKAGE"