+++ R/plotShiny.R
@@ -11,10 +11,10 @@
 #' 
 #' @param wldf an object of class \linkS4class{WaterLevelDataFrame}.
 #' @param add_flys \code{logical} determining whether the used 
-#'   \href{http://www.bafg.de/DE/08_Ref/M2/03_Fliessgewmod/01_FLYS/flys_node.html}{FLYS3}
+#'   \href{https://www.bafg.de/DE/5_Informiert/1_Portale_Dienste/FLYS/flys_node.html}{FLYS3}
 #'   water levels should be plotted.
 #' @param add_flys_labels \code{logical} determining whether the used 
-#'   \href{http://www.bafg.de/DE/08_Ref/M2/03_Fliessgewmod/01_FLYS/flys_node.html}{FLYS3}
+#'   \href{https://www.bafg.de/DE/5_Informiert/1_Portale_Dienste/FLYS/flys_node.html}{FLYS3}
 #'   water levels should be labelled.
 #' @param add_weighting \code{logical} determining whether the weighting of
 #'   gauging data at the gauging stations should be labelled.