+++ R/updateGaugingData.R
@@ -42,11 +42,9 @@ updateGaugingData <- function(x) {
     
     if (!file.exists(x)) {
         tryCatch({
-            utils::download.file(url, x, quiet = TRUE, method = "curl")
+            .download_bfg(url, x)
             invisible(TRUE)
         }, error = function(e){
-            message(paste0("It was not possible to update the gauging data. Tr",
-                           "y again later!"))
             invisible(FALSE)
             }
         )
@@ -54,11 +52,9 @@ updateGaugingData <- function(x) {
         file_mtime <- file.info(x)$mtime
         if (file_mtime < Sys.time() - 8 * 24 * 60 * 60) {
             tryCatch({
-                utils::download.file(url, x, quiet = TRUE, method = "curl")
+                .download_bfg(url, x)
                 invisible(TRUE)
             }, error = function(e){
-                message(paste0("It was not possible to update the gauging data",
-                               ". Try again later!"))
                 invisible(FALSE)
             }
             )
@@ -66,4 +62,39 @@ updateGaugingData <- function(x) {
             invisible(FALSE)
         }
     }
+}
+
+.download_bfg <- function(x, file) {
+    # check internet
+    if (!curl::has_internet()) {
+        stop(paste0("Dataset provided by hyd1d.bafg.de is unavailable without ",
+                    "internet."), call. = FALSE)
+    }
+    
+    # assemble request
+    req <- httr2::request(x)
+    req <- httr2::req_method(req, "GET")
+    req <- httr2::req_retry(req, max_tries = 5L)
+    req <- httr2::req_timeout(req, seconds = options()$timeout)
+    req <- httr2::req_error(req, is_error = \(resp) FALSE)
+    
+    # perform the request
+    resp <- httr2::req_perform(req, path = file, verbosity = 0)
+    
+    # handle errors
+    status_code <- as.character(resp$status_code)
+    if (startsWith(status_code, "4")) {
+        mess <- paste0("The request to hyd1d.bafg.de returned a status code of",
+                       ":\n   '", status_code, " - ", resp_status_desc(resp),
+                       "'\nPlease adjust your request accordingly:\n   url: ",
+                       x)
+        stop(mess, call. = FALSE)
+    }
+    if (startsWith(status_code, "5")) {
+        mess <- paste0("The request to hyd1d.bafg.de returned a status code of",
+                       ":\n   '", status_code, " - ", resp_status_desc(resp),
+                       "'\nPlease try again later.")
+        stop(mess, call. = FALSE)
+    }
+    return(TRUE)
 }